server/requirements.txt
asgiref
autopep8
cachetools
certifi
chardet
coreapi
coreschema
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-yasg
Faker
future
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
gunicorn
httplib2
idna
inflection
itypes
Jinja2
MarkupSafe
oauthlib
packaging
protobuf
pyasn1
pyasn1-modules
pycodestyle
PyJWT
pyparsing
python-dateutil
python-twitter
pytz
requests
requests-oauthlib
rsa
ruamel.yaml
ruamel.yaml.clib
six
sqlparse
text-unidecode
toml
uritemplate
urllib3
psycopg2-binary
Pillow
django-filter